Reduce internal testing in arcam_fmj tests
